test(flip-api): split presigned-URL policy tests by source file
Follow the project convention `test_<source_file>.py`:

* `tests/unit/utils/test_s3_client.py` — covers
  `flip_api/utils/s3_client.py`: the URL-redaction logging contract on
  `get_put_presigned_url`, the silent TTL cap at the security ceiling,
  and the default 600 s TTL.
* `tests/unit/file_services/test_presigned_url_for_upload.py` —
  covers `flip_api/file_services/presigned_url_for_upload.py`: success
  path, `PRE_SIGNED_URL` env-override path, and short-circuit on
  `fileName` validation (asserts S3 client is never instantiated when
  the schema rejects the body).
* `tests/unit/file_services/test_retrieve_federated_results.py` —
  appends the URL-list logging-policy test to the existing module that
  already covers this route.
* `tests/unit/domain/schemas/test_file.py` — covers
  `flip_api/domain/schemas/file.py`: the parametrised
  `UploadFileBody.fileName` validator cases.

Removes the now-redundant `test_presigned_url_logging_policy.py`
created in the previous commit.

Retest: 890 unit tests pass; the four split files contribute 36
checks (3 s3_client + 3 upload route + 1 federated-results + 29
schema = 36); ruff + mypy clean.
